fix: add node affinity based on resource label (#152)
By design, the resource.labelselector should be used to add node
affinity to make sure workload run in the provisioned nodes.
The workload label selector should use a different one. This makes it
possible for two workspaces to share the same nodes.

Also, remove pod anti-affinity rule which is unnecessary. The Pod GPU
request makes sure the GPU node is not overcommitted.

Co-authored-by: guofei <guofei@microsoft.com>