docs: move installation steps to a separate file. (#226)
The installation steps are very long. It is better to move it to a
separate file. Or the installation steps are slightly adjust.

The following statement is added:
```
Note that if you have installed another node provisioning controller that supports Karpenter-core APIs, the following steps for installing `gpu-provisioner` can be skipped.
```

Co-authored-by: guofei <guofei@microsoft.com>