feat: Update CRD to add Volume and ConfigTemplate (#356)
**Reason for Change**:
This change removes hostPath field in DataSource and DataDestination
struct. Instead, a standard K8s VolumeSource API is used to specify the
volumes used by the tuning Job. This will provide the highest
flexibility for users to load/store the data. HostPath field is hard to
use in practice.

This change also renames the Config to ConfigTemplate because the final
configmap used by the tuning job will be generated by the controller
based on the template. The new name is clearer.

**Notes for Reviewers**:
We will revisit how to do validation check for the Volume field because
it is a fairly complicated API.