fix: add watcher for job and avoid unnecessary reconcile (#527)
This PR addresses two problems:

1) We should add watcher for tunning jobs. Otherwise, if a tunning job
is deleted before it is done, the controller will not recreate it.
2) Currently, the gpu-provisioner will update the machine object with
heartbeat information every two minutes. This will trigger a workspace
reconcile. In practice, if the machine keeps in the ready state, we
don't need to reconcile the workspace. Add an additional machine event
filter to avoid unnecessary reconciles.