fix: set job's backoffLimit to 0 and report job's active/ready status in workspace (#575)
For Job workload, the backoffLimit specifies the number of retries
before marking this job failed. Defaults to 6. In Kaito finetuning,
there are not many benefits to recreate pods 6 times if the tuning
fails, because the failures usually require user interventions.

This change also reports the job's ready pod count in workspace CR so
that users do not need to frequently check the job object for status
check. When tuning job is working, the ready count is 1. If the tuning
container completes but the docker sidecar containers fails with
infinite retry, the ready count becomes 0.

Add another tip in the troubleshoot guide.