Avoid creating GPU provisioner in the default namespace (#706)
**Reason for Change**:
<!-- What does this PR improve or fix in Kaito? Why is it needed? -->
I discussed an issue that the default installation of Kaito failed to
work as it could create the GPU provisioner in a default name space and
customers need to manually change the namespace to "default" to ensure
the GPU provisioner pod is in running status

**Requirements**
Fix the document for the helm install command. The command is validated
in the latest AKS cluster and discussed with Fei.
- [ ] added unit tests and e2e tests (if applicable).

**Issue Fixed**:
<!-- If this PR fixes GitHub issue 4321, add "Fixes #4321" to the next
line. -->

**Notes for Reviewers**:

Co-authored-by: HJ Hang <hjhang@HJs-MacBook-Pro.local>