feat: [SKU modularization] AWS chart changes (#710)
**Reason for Change**:
<!-- What does this PR improve or fix in Kaito? Why is it needed? -->
Updating kaito charts for aws integration:

- add clusterrole permission for ec2nodeclasses resource
- add new env variable in deployment CLUSTER_NAME as this environment is
needed to create EC2 Node Class
- update nodeSelectorTerm for nvidia-device-plugin daemonset based on
the cloud provider
- add AWS specific nodeSelectorRequirement for nodeClaim obj
- add InstanceStorePolicyRAID0 to EC2 node class to share node's
ephermeral storage among pods that request it. Additional information
[here](https://karpenter.sh/v1.0/concepts/nodeclasses/#specinstancestorepolicy)
- fixing a bug on nodeClaim create (kubeClient.Create) - if the
nodeClaim fails to get created, we retry to create it again. However, on
kubeClient.Create, certain metadata are set like resourceVersion which
will be passed along in the retires if we don't pass the DeepCopy().
This metadata being populated will throw an error from KubeClient
`resourceVersion should not be set on objects to be created`